requirements.txt
streamlit
plotly
azure-storage-blob
